build empty skeleton
