+++ antarest/study/business/model/user_model.py
@@ -1,0 +1,28 @@
+# Copyright (c) 2025, RTE (https://www.rte-france.com)
+#
+# See AUTHORS.txt
+#
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+#
+# SPDX-License-Identifier: MPL-2.0
+#
+# This file is part of the Antares project.
+from enum import StrEnum
+
+from antarest.core.serde import AntaresBaseModel
+
+
+class ResourceType(StrEnum):
+    FILE = "file"
+    FOLDER = "folder"
+
+
+class CreateUserResourceData(AntaresBaseModel):
+    path: str
+    resource_type: ResourceType
+
+
+class RemoveUserResourceData(AntaresBaseModel):
+    path: str
