+++ antarest/study/dao/api/user_resources_dao.py
@@ -1,0 +1,29 @@
+# Copyright (c) 2025, RTE (https://www.rte-france.com)
+#
+# See AUTHORS.txt
+#
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+#
+# SPDX-License-Identifier: MPL-2.0
+#
+# This file is part of the Antares project.
+from abc import ABC, abstractmethod
+from pathlib import PurePosixPath
+
+from antarest.study.business.model.user_model import ResourceType
+
+
+class ReadOnlyUserResourcesDao(ABC):
+    pass
+
+
+class UserResourcesDao(ReadOnlyUserResourcesDao):
+    @abstractmethod
+    def save_user_resource(self, resource_type: ResourceType, resource_path: PurePosixPath) -> None:
+        raise NotImplementedError()
+
+    @abstractmethod
+    def delete_user_resource(self, resource_path: PurePosixPath) -> None:
+        raise NotImplementedError()
