+++ antarest/study/dao/file/file_study_dao.py
@@ -27,6 +27,7 @@ from antarest.study.dao.file.file_study_st_storage_dao import FileStudySTStorage
 from antarest.study.dao.file.file_study_thematic_trimming_dao import FileStudyThematicTrimmingDao
 from antarest.study.dao.file.file_study_thermal_dao import FileStudyThermalDao
 from antarest.study.dao.file.file_study_timseries_config_dao import FileStudyTimeSeriesConfigDao
+from antarest.study.dao.file.file_study_user_resources_dao import FileStudyUserResourceDao
 from antarest.study.dao.file.file_study_xpansion_dao import FileStudyXpansionDao
 from antarest.study.storage.rawstudy.model.filesystem.factory import FileStudy
 
@@ -47,6 +48,7 @@ class FileStudyTreeDao(
     FileStudyAdequacyPatchParametersDao,
     FileStudyTimeSeriesConfigDao,
     FileStudyLayerDao,
+    FileStudyUserResourceDao,
 ):
     """
     Implementation of study DAO over the simulator input format.
