+++ antarest/study/storage/variantstudy/model/command/remove_user_resource.py
@@ -15,7 +15,7 @@ from typing import List, Optional, cast
 from typing_extensions import override
 
 from antarest.core.exceptions import ChildNotFoundError
-from antarest.core.serde import AntaresBaseModel
+from antarest.study.business.model.user_model import RemoveUserResourceData
 from antarest.study.storage.rawstudy.model.filesystem.factory import FileStudy
 from antarest.study.storage.rawstudy.model.filesystem.root.user.user import User
 from antarest.study.storage.variantstudy.model.command.common import (
@@ -28,10 +28,6 @@ from antarest.study.storage.variantstudy.model.command.common import (
 from antarest.study.storage.variantstudy.model.command.icommand import ICommand
 from antarest.study.storage.variantstudy.model.command_listener.command_listener import ICommandListener
 from antarest.study.storage.variantstudy.model.model import CommandDTO
-
-
-class RemoveUserResourceData(AntaresBaseModel):
-    path: str
 
 
 class RemoveUserResource(ICommand):
